refactor: constants를 위한 package 생성
