+++ src/main/java/lotto/constant/Rank.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.constant;
 
 import java.util.Optional;
