+++ src/main/java/lotto/domain/LottoScanner.java
@@ -5,7 +5,7 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Stream;
-import lotto.Rank;
+import lotto.constant.Rank;
 
 public class LottoScanner {
     public LottoScanner() {
