+++ src/test/java/lotto/LottoScannerTest.java
@@ -2,9 +2,8 @@ package lotto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
+import lotto.constant.Rank;
 import lotto.domain.Lotto;
 import lotto.domain.LottoScanner;
 import lotto.domain.Lottos;
